git-svn-id: https://svn.code.sf.net/p/pinmame/code/branches/dll@4479 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/cheat.c
@@ -8005,7 +8005,7 @@ static void SetSearchRegionDefaultName(SearchRegion * region)
 				if(	(handlerAddress >= ((size_t)MWA_BANK1)) &&
 					(handlerAddress <= ((size_t)MWA_BANK24)))
 				{
-					sprintf(desc, "BANK%.2d", (handlerAddress - ((size_t)MWA_BANK1)) + 1);
+					sprintf(desc, "BANK%.2d", (unsigned int)(handlerAddress - ((size_t)MWA_BANK1)) + 1);
 				}
 				else
 					if(handlerAddress == (size_t)MWA_NOP)
